requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Filtering
django-filter==23.5

# Nested Routers
drf-nested-routers==0.93.5

# Payments
stripe==8.4.0

# Email
resend==2.0.0

# File Storage
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Async Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0

# Utilities
Pillow==10.2.0
openpyxl==3.1.2

# Testing
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
